+++ node/directives/validateAdminUserAccess.ts
@@ -43,7 +43,8 @@ export class ValidateAdminUserAccess extends SchemaDirectiveVisitor {
 
       const { hasAdminToken, hasValidAdminToken } = await validateAdminToken(
         context,
-        adminUserAuthToken as string
+        adminUserAuthToken as string,
+        metricFields
       )
 
       // add admin token metrics
@@ -69,7 +70,7 @@ export class ValidateAdminUserAccess extends SchemaDirectiveVisitor {
 
       // If there's no valid admin token on context, search for it on header
       const { hasAdminTokenOnHeader, hasValidAdminTokenOnHeader } =
-        await validateAdminTokenOnHeader(context)
+        await validateAdminTokenOnHeader(context, metricFields)
 
       // add admin header token metrics
       metricFields = {
@@ -92,7 +93,10 @@ export class ValidateAdminUserAccess extends SchemaDirectiveVisitor {
         return resolve(root, args, context, info)
       }
 
-      const { hasApiToken, hasValidApiToken } = await validateApiToken(context)
+      const { hasApiToken, hasValidApiToken } = await validateApiToken(
+        context,
+        metricFields
+      )
 
       // add API token metrics
       metricFields = {